backend/requirements.txt
flask>=3.0
flask-cors>=4.0
flask-sqlalchemy>=3.1
flask-limiter>=3.5
psycopg2-binary>=2.9
python-dotenv>=1.0
pyjwt>=2.8
openai>=1.12
numpy>=1.26
gunicorn>=21.2
bcrypt>=4.1
